Separating logic for SB and attachable explosives
