+++ UnityProject/Assets/Prefabs/Items/Weapons/Explosive/SyndicateBomb.prefab
@@ -186,8 +186,7 @@ PrefabInstance:
       propertyPath: ItemAttributesV2
       value: 
       objectReference: {fileID: 5052867042963084228}
-    m_RemovedComponents:
-    - {fileID: 5745720402341052027, guid: 0a45eeafcc7b5164bb139db7bf25c0a8, type: 3}
+    m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 0a45eeafcc7b5164bb139db7bf25c0a8, type: 3}
 --- !u!1 &747172844020729843 stripped
 GameObject:
@@ -273,6 +272,7 @@ MonoBehaviour:
       m_AssetGUID: 
       m_SubObjectName: 
       m_SubObjectType: 
+      m_EditorAssetChanged: 0
   soundDelayTime: 0
   soundMinimumPitchVariance: 1
   soundMaximumPitchVariance: 1
@@ -303,17 +303,3 @@ MonoBehaviour:
   initialPassable: 0
   initialCrawlPassable: 0
   passableExclusionsToThis: []
---- !u!114 &3021885719685710939
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 747172844020729843}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 6d7277dcc993463cb6c7f79b663f36fc, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  NetTabType: 46
-  aiInteractable: 0
